requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
amqp==5.3.1
    # via
    #   -r requirements/base.txt
    #   kombu
appdirs==1.4.4
    # via
    #   -r requirements/base.txt
    #   fs
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
astroid==3.3.9
    # via
    #   pylint
    #   pylint-celery
billiard==4.2.1
    # via
    #   -r requirements/base.txt
    #   celery
celery==5.5.2
    # via
    #   -r requirements/base.txt
    #   edx-celeryutils
certifi==2025.4.26
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   pynacl
charset-normalizer==3.4.2
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.8
    # via
    #   -r requirements/base.txt
    #   celery
    #   click-didyoumean
    #   click-log
    #   click-plugins
    #   click-repl
    #   code-annotations
    #   edx-django-utils
    #   edx-lint
click-didyoumean==0.3.1
    # via
    #   -r requirements/base.txt
    #   celery
click-log==0.4.0
    # via edx-lint
click-plugins==1.1.1
    # via
    #   -r requirements/base.txt
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements/base.txt
    #   celery
code-annotations==2.3.0
    # via edx-lint
coverage[toml]==7.8.0
    # via pytest-cov
dill==0.4.0
    # via pylint
    # via
    #   -c requirements/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-model-utils
    #   django-waffle
    #   djangorestframework
    #   edx-bulk-grades
    #   edx-celeryutils
    #   edx-django-utils
    #   jsonfield
    #   super-csv
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   super-csv
django-model-utils==5.0.0
    # via
    #   -r requirements/base.txt
    #   edx-bulk-grades
    #   edx-celeryutils
    #   super-csv
django-waffle==4.2.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-statici18n==2.6.0
    # via -r requirements/dev.in
djangorestframework==3.16.0
    # via
    #   -r requirements/base.txt
    #   super-csv
dnspython==2.7.0
    # via
    #   -r requirements/base.txt
    #   pymongo
edx-bulk-grades==1.2.0
    # via -r requirements/base.txt
edx-celeryutils==1.3.0
    # via
    #   -r requirements/base.txt
    #   super-csv
edx-django-utils==7.4.0
    # via
    #   -r requirements/base.txt
    #   super-csv
edx-lint==5.6.0
    # via -r requirements/test.in
edx-opaque-keys==3.0.0
    # via
    #   -r requirements/base.txt
    #   edx-bulk-grades
fs==2.4.16
    # via
    #   -r requirements/base.txt
    #   xblock
idna==3.10
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via pylint
jinja2==3.1.6
    # via code-annotations
jsonfield==3.1.0
    # via
    #   -r requirements/base.txt
    #   edx-celeryutils
kombu==5.5.3
    # via
    #   -r requirements/base.txt
    #   celery
lxml==5.4.0
    # via
    #   -r requirements/base.txt
    #   xblock
mako==1.3.10
    # via
    #   -r requirements/base.txt
    #   xblock
markdown==3.8
    # via -r requirements/base.txt
markupsafe==3.0.2
    # via
    #   -r requirements/base.txt
    #   jinja2
    #   mako
    #   xblock
mccabe==0.7.0
    # via pylint
newrelic==10.11.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
packaging==25.0
    # via pytest
path==17.1.0
    # via
    #   -r requirements/base.txt
    #   path-py
path-py==12.5.0
    # via -r requirements/base.txt
pbr==6.1.1
    # via
    #   -r requirements/base.txt
    #   stevedore
platformdirs==4.3.7
    # via pylint
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.51
    # via
    #   -r requirements/base.txt
    #   click-repl
psutil==7.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pylint==3.3.7
    # via
    #   edx-lint
    #   pylint-celery
    #   pylint-django
    #   pylint-plugin-utils
pylint-celery==0.3
    # via edx-lint
pylint-django==2.6.1
    # via edx-lint
pylint-plugin-utils==0.8.2
    # via
    #   pylint-celery
    #   pylint-django
pymongo==4.12.1
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==8.3.5
    # via pytest-cov
pytest-cov==6.1.1
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.txt
    #   celery
    #   xblock
python-slugify==8.0.4
    # via code-annotations
pytz==2025.2
    # via
    #   -r requirements/base.txt
    #   xblock
pyyaml==6.0.2
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   xblock
requests==2.32.3
    # via
    #   -r requirements/base.txt
    #   edx-bulk-grades
    #   slumber
simplejson==3.20.1
    # via
    #   -r requirements/base.txt
    #   super-csv
    #   xblock
six==1.17.0
    # via
    #   -r requirements/base.txt
    #   edx-lint
    #   fs
    #   python-dateutil
slumber==0.7.1
    # via
    #   -r requirements/base.txt
    #   edx-bulk-grades
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
super-csv==4.1.0
    # via
    #   -r requirements/base.txt
    #   edx-bulk-grades
text-unidecode==1.3
    # via python-slugify
tomlkit==0.13.2
    # via pylint
typing-extensions==4.13.2
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
tzdata==2025.2
    # via
    #   -r requirements/base.txt
    #   kombu
urllib3==2.2.3
    # via
    #   -c requirements/common_constraints.txt
    #   -r requirements/base.txt
    #   requests
vine==5.1.0
    # via
    #   -r requirements/base.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   -r requirements/base.txt
    #   prompt-toolkit
web-fragments==3.0.0
    # via
    #   -r requirements/base.txt
    #   xblock
webob==1.8.9
    # via
    #   -r requirements/base.txt
    #   xblock
xblock==5.2.0
    # via -r requirements/base.txt

# The following packages are considered to be unsafe in a requirements file:
# setuptools
